server/requirements.txt
fastapi
uvicorn
python-socketio[asyncio_client]